backend/requirements.txt
asgiref==3.4.1
Django==3.2.6
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
psycopg2==2.9.3
django-cors-headers==3.11.0